+++ src/main/java/org/eclipse/keyple/core/util/logging/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Logging abstraction layer.
+ *
+ * @since 2.5.0
+ */
+package org.eclipse.keyple.core.util.logging;