+++ src/test/java/org/eclipse/keyple/core/util/logging/Slf4jLoggerProviderTest.java
@@ -1,0 +1,27 @@
+/* **************************************************************************************
+ * Copyright (c) 2026 Calypso Networks Association https://calypsonet.org/
+ *
+ * See the NOTICE file(s) distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ************************************************************************************** */
+package org.eclipse.keyple.core.util.logging;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.Test;
+
+public class Slf4jLoggerProviderTest {
+
+  @Test
+  public void shouldReturnSlf4jLogger() {
+    Slf4jLoggerProvider provider = new Slf4jLoggerProvider();
+    Logger logger = provider.getLogger("test.logger");
+
+    assertThat(logger).isInstanceOf(Slf4jLogger.class);
+  }
+}